pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
name = "llarmy"
version = "0.0.1"
description = "A plethora of small llama-index based agents and tools to create agentic systems"
readme = "README.md"
requires-python = ">=3.9,<4.0"
license = {text = "MIT"}
authors = [{email = "rgonzalezp1115@gmail.com", name = "Ricky Gonzalez Penuela"}]
maintainers = [
    {email = "rgonzalezp1115@gmail.com", name = "Ricky Gonzalez Penuela"},
]
keywords = [
    "LLM",
    "NLP",
    "RAG",
    "data",
    "devtools",
    "index",
    "retrieval",
    "agents",
    "agents_tools"
]
classifiers = [

    "Programming Language :: Python :: 3.9",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
]

[project.urls]
Repository = "https://github.com/llarmy"

[project.optional-dependencies]
dev = [
    "black[jupyter]>=23.7.0,<=24.3.0",
    "codespell[toml]>=v2.2.6",
    "mypy==1.11.0",
    "pre-commit==3.2.0",
    "pylint==2.15.10",
    "pytest-asyncio>=0.23.7",
    "pytest-mock>=3.14.0",
    "pytest>=8.2.1",
    "python-dotenv>=1.0.0",
    "ruff==0.11.11",
    "twine>=4.0.0",
    "types-Deprecated>=0.1.0",
    "types-PyYAML>=6.0.12.12,<7",
    "types-protobuf>=4.24.0.4,<5",
    "types-redis==4.5.5.0",
    "types-requests==2.28.11.8",
    "types-setuptools==67.1.0.0",
    "tomli>=2.2.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/llarmy"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.mypy]
disallow_untyped_defs = true
exclude = ["build", "venv"]
explicit_package_bases = true
ignore_missing_imports = true
mypy_path = "src"
namespace_packages = true
python_version = "3.9"

[tool.ruff]
target-version = "py39"
line-length = 88
exclude = [
    "build",
    "venv",
]

[tool.ruff.lint]
ignore = [
    "COM812",  # Too aggressive
    "D212",  # Using D213
    "D417",  # Too aggressive
    "F541",  # Messes with prompts.py
    "RUF100",  # Allow blanket noqa
    "TC002",
    "UP",  # Remove when we drop Python 3.9
    "PT001",
    "E501",  # Use best judgement for line-length
    "E402",  # Annoying, use best judgement
    "PYI063",
    "ANN204",  # this is annoying
    "D401",  # I disagree
    "D404",
]
select = [
    "ANN204",
    "B009",
    "B010",
    "B011",
    "B013",
    "B014",
    "C4",
    "COM812",
    "COM819",
    "D201",
    "D202",
    "D204",
    "D207",
    "D208",
    "D209",
    "D211",
    "D213",
    "D214",
    "D215",
    "D3",
    "D4",
    "E",
    "EXE004",
    "F401",
    "F504",
    "F541",
    "F632",
    "FLY",
    "G010",
    "I002",
    "PERF1",
    "PIE790",
    "PIE794",
    "PIE808",
    "PIE810",
    "PLC0414",
    "PLE2510",
    "PLE2512",
    "PLE2513",
    "PLE2514",
    "PLE2515",
    "PLR1711",
    "PT001",
    "PT003",
    "PT006",
    "PT02",
    "PTH201",
    "PYI",
    "Q",
    "RET501",
    "RET502",
    "RET503",
    "RET504",
    "RSE",
    "RUF005",
    "RUF010",
    "RUF015",
    "RUF1",
    "SIM101",
    "SIM103",
    "SIM109",
    "SIM118",
    "SIM2",
    "SIM300",
    "SIM9",
    "TC005",
    "TD006",
    "TID",
    "TRY201",
    "W",
]
unfixable = [
    "ERA001",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.codespell]
check-filenames = true
check-hidden = true
skip = "./build,./venv,*.csv,*.html,*.json,*.jsonl,*.pdf,*.txt,*.ipynb"

[tool.tomlsort]
all = false
in_place = true
spaces_before_inline_comment = 2  # Match Python PEP 8
spaces_indent_inline_array = 4  # Match Python PEP 8
trailing_comma_inline_array = true
